fix(externalclock): fix deadlock when sending time on unstarted Clock
Calling `SetTimestamp` on a clock will block until the `Run` goroutine
has read on channel and acked on another channel. If the Clock isn't
started (or has been closed) calling `SetTimestamp` blocks
forever (unless clock is started later).

This commit fixes this issue by setting time and signaling tickers in
the calling goroutine (as opposed to `Run` goroutine).

This change also means that `Run` doesn't do anything, so it is marked
as deprecated.