fix: remove data race when resetting ticker
Currently there is a data race in the Ticker that is triggered when
Ticker.Reset and Ticker.IsDurationReached are called concurrently.
IsDurationReached is called by the clock as part of Clock.SetTimestamp and
Ticker.Reset is called by user-code.

This commit fixes the data race by protecting `duration` field of Ticker
with the mutex. This also forces inlining of some calls in both methods
as the called methods will also attempt to lock the mutex.